+++ linker.h
@@ -1,4 +1,4 @@
 #pragma once
-void transposeExt(float input[5][5]);
-void multExt(float inputArr[5][5], int multiplier);
-void inverseExt(float inputArr[5][5]);
+void transposeExt(double input[5][5]);
+void multExt(double inputArr[5][5], int multiplier);
+void inverseExt(double inputArr[5][5]);
